delegate dap.ps domain management to Route53
We have to use Route53 because it provides the option to use an
ALIAS type record which works for apex records and can point to
a CloudFront distribution in a dynamic way.

Without this we'd have to use A records which are static unlike
a CNAME, which would eventually become obsolete and take down
the site when they do.

Details: https://github.com/dap-ps/infra-dapps/issues/18

Signed-off-by: Jakub Sokołowski <jakub@status.im>